Move parsec._parsec typing folder to avoid error from importlib.resources.contents
